+++ packages/types/src/chains/index.ts
@@ -1,5 +1,5 @@
 import { Chain, ChainId, StagenetChain } from "./_enums";
-import { CosmosChainConfigs } from "./cosmos";
+import { CosmosChainConfigs, StagenetMAYAConfig, StagenetTHORConfig } from "./cosmos";
 import { EVMChainConfigs } from "./evm";
 import { OtherChainConfigs } from "./others";
 import { SubstrateChainConfigs } from "./substrate";
@@ -46,21 +46,59 @@ export function getChainConfig<T extends keyof ChainConfigMap>(chainOrChainId: T
   return (chainConfig || {}) as ChainConfigMap[T];
 }
 
-const { chainIdToChain, chainToBaseDecimal, chainToBlockTime } = AllChains.reduce(
+const { chainIdToChain, chainToBaseDecimal, chainToBlockTime, explorerUrls, rpcUrls } = AllChains.reduce(
   (acc, chain) => {
-    const { chainId, baseDecimal, blockTime } = getChainConfig(chain);
+    const { chainId, baseDecimal, blockTime, explorerUrl, rpcUrls } = getChainConfig(chain);
 
     acc.chainIdToChain[chainId] = chain;
     acc.chainToBaseDecimal[chain] = baseDecimal;
     acc.chainToBlockTime[chain] = blockTime;
+    acc.explorerUrls[chain] = explorerUrl;
+    acc.rpcUrls[chain] = rpcUrls[0] || "";
     return acc;
   },
-  {} as {
+  {
+    rpcUrls: {
+      [StagenetChain.Maya]: StagenetMAYAConfig.rpcUrls[0],
+      [StagenetChain.THORChain]: StagenetTHORConfig.rpcUrls[0],
+    },
+  } as {
     chainIdToChain: Record<ChainId, Chain>;
     chainToBaseDecimal: Record<Chain, number>;
     chainToBlockTime: Record<Chain, number>;
+    explorerUrls: Record<Chain, string>;
+    rpcUrls: Record<Chain | StagenetChain, string>;
   },
 );
+
+/**
+ * @deprecated
+ */
+export const NODE_URLS = {
+  [Chain.THORChain]: "https://thornode.ninerealms.com",
+  [Chain.Maya]: "https://mayanode.mayachain.info",
+  [StagenetChain.THORChain]: "https://stagenet-thornode.ninerealms.com",
+  [StagenetChain.Maya]: "https://stagenet.mayanode.mayachain.info",
+};
+
+/**
+ * @deprecated use getChainConfig instead
+ * @example
+ * ```diff
+ * -const rpcUrl = RPC_URLS[Chain.Ethereum];
+ * +const { rpcUrls: [rpcUrl] } = getChainConfig(Chain.Ethereum);
+ * ```
+ */
+export const RPC_URLS: Record<Chain | StagenetChain, string> = rpcUrls;
+
+/**
+ * @deprecated use getChainConfig instead
+ * @example
+ * ```diff
+ * -const explorerUrl = EXPLORER_URLS[Chain.Ethereum];
+ * +const { blockExplorerUrl } = getChainConfig(Chain.Ethereum);
+ */
+export const EXPLORER_URLS: Record<Chain, string> = explorerUrls;
 
 /**
  *