+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @swapkit/ui
 
+## 0.1.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/browser@4.0.24
+
+## 0.1.23
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/browser@4.0.23
+
 ## 0.1.22
 
 ### Patch Changes